Intermediate class between `Market` and `GeminiMarket`
